add file_patch_inplace(), fix file_patch when src and dst are the same file

+++ bsdiff4/format.py
@@ -97,12 +97,32 @@ def patch(src_bytes, patch_bytes):
     return core.patch(src_bytes, *read_patch(StringIO(patch_bytes)))
 
 
+def file_patch_inplace(path, patch_path):
+    """file_patch_inplace(path, patch_path)
+
+    Apply the BSDIFF4-format file patch_path to the file 'path' in place.
+    """
+    fi = open(patch_path, 'rb')
+    f = open(path, 'r+b')
+    data = f.read()
+    f.seek(0)
+    f.write(core.patch(data, *read_patch(fi)))
+    f.close()
+    fi.close()
+
+
 def file_patch(src_path, dst_path, patch_path):
     """file_patch(src_path, dst_path, patch_path)
 
     Apply the BSDIFF4-format file patch_path to the file src_path and
     write the result to the file dst_path.
     """
+    from os.path import isfile, samefile
+
+    if isfile(dst_path) and samefile(dst_path, src_path):
+        file_patch_inplace(src_path, patch_path)
+        return
+
     fi = open(patch_path, 'rb')
     fo = open(dst_path, 'wb')
     fo.write(core.patch(read_data(src_path), *read_patch(fi)))
